Move HookHelper to contao/test-case to fix autoloading
