+++ test-case/src/HookHelper.php
@@ -10,7 +10,7 @@ declare(strict_types=1);
  * @license LGPL-3.0-or-later
  */
 
-namespace Contao\CoreBundle\Tests\Fixtures\Helper;
+namespace Contao\TestCase;
 
 class HookHelper
 {
